refactor(api): claude_sessions 동적 라우트를 sessions·activity 모듈로 추출
/{session_id} 계열 7개를 sessions.py로, /{session_id}/activity·/tasks 3개를
activity.py로 이동. 핸들러 본문·경로·함수명 무변경 (원문 구간 diff 로
바이트 동일 확인).

동적 라우트를 **먼저** 빼내는 이유: include_router 는 라우트를 리스트 끝에
붙이므로, 구체 경로를 먼저 추출하면 _legacy 에 남은 GET/DELETE /{session_id}
가 그것을 가려버린다. 파라미터 경로 모듈을 마지막에 include 하는 불변식을
첫 추출부터 성립시킨다.

트랜스크립트 줄 수 캐시(LineCountCacheEntry·TranscriptLineCountCache)는
get_session_transcript 단독 소비자라 sessions.py 로 함께 이동했다.

라우트 테이블 characterization 테스트 통과 — HTTP 표면 25개 동일.
